+++ lib/src/multi_split_view.dart
@@ -10,7 +10,6 @@ import 'package:multi_split_view/src/internal/layout_delegate.dart';
 import 'package:multi_split_view/src/policies.dart';
 import 'package:multi_split_view/src/theme_data.dart';
 import 'package:multi_split_view/src/theme_widget.dart';
-import 'package:multi_split_view/src/typedefs.dart';
 
 /// A widget to provides horizontal or vertical multiple split view.
 class MultiSplitView extends StatefulWidget {
@@ -22,107 +21,28 @@ class MultiSplitView extends StatefulWidget {
   /// The [children] argument is required.
   const MultiSplitView(
       {Key? key,
-      Axis axis = MultiSplitView.defaultAxis,
-      required List<Widget> children,
-      MultiSplitViewController? controller,
-      DividerBuilder? dividerBuilder,
-      OnDividerDragUpdate? onDividerDragUpdate,
-      DividerTapCallback? onDividerTap,
-      DividerTapCallback? onDividerDoubleTap,
-      bool resizable = true,
-      bool antiAliasingWorkaround = false,
-      bool pushDividers = false,
-      List<Area>? initialAreas,
-      SizeOverflowPolicy sizeOverflowPolicy = SizeOverflowPolicy.shrinkLast,
-      SizeUnderflowPolicy sizeUnderflowPolicy = SizeUnderflowPolicy.stretchLast,
-      double fallbackWidth = 500,
-      double fallbackHeight = 500})
-      : this._(
-            key: key,
-            axis: axis,
-            children: children,
-            controller: controller,
-            dividerBuilder: dividerBuilder,
-            onDividerDragUpdate: onDividerDragUpdate,
-            onDividerTap: onDividerTap,
-            onDividerDoubleTap: onDividerDoubleTap,
-            resizable: resizable,
-            antiAliasingWorkaround: antiAliasingWorkaround,
-            count: null,
-            widgetBuilder: null,
-            pushDividers: pushDividers,
-            initialAreas: initialAreas,
-            sizeOverflowPolicy: sizeOverflowPolicy,
-            sizeUnderflowPolicy: sizeUnderflowPolicy,
-            fallbackWidth: fallbackWidth,
-            fallbackHeight: fallbackHeight);
-
-  /// Creates an [MultiSplitView].
-  ///
-  /// The default value for [axis] argument is [Axis.horizontal].
-  const MultiSplitView.builder(
-      {Key? key,
-      Axis axis = MultiSplitView.defaultAxis,
-      MultiSplitViewController? controller,
-      required int count,
-      required AreaWidgetBuilder widgetBuilder,
-      DividerBuilder? dividerBuilder,
-      OnDividerDragUpdate? onDividerDragUpdate,
-      DividerTapCallback? onDividerTap,
-      DividerTapCallback? onDividerDoubleTap,
-      bool resizable = true,
-      bool antiAliasingWorkaround = false,
-      bool pushDividers = false,
-      List<Area>? initialAreas,
-      SizeOverflowPolicy sizeOverflowPolicy = SizeOverflowPolicy.shrinkLast,
-      SizeUnderflowPolicy sizeUnderflowPolicy = SizeUnderflowPolicy.stretchLast,
-      double fallbackWidth = 500,
-      double fallbackHeight = 500})
-      : this._(
-            key: key,
-            axis: axis,
-            children: null,
-            controller: controller,
-            dividerBuilder: dividerBuilder,
-            onDividerDragUpdate: onDividerDragUpdate,
-            onDividerTap: onDividerTap,
-            onDividerDoubleTap: onDividerDoubleTap,
-            resizable: resizable,
-            antiAliasingWorkaround: antiAliasingWorkaround,
-            count: count,
-            widgetBuilder: widgetBuilder,
-            initialAreas: initialAreas,
-            pushDividers: pushDividers,
-            sizeOverflowPolicy: sizeOverflowPolicy,
-            sizeUnderflowPolicy: sizeUnderflowPolicy,
-            fallbackWidth: fallbackWidth,
-            fallbackHeight: fallbackHeight);
-
-  const MultiSplitView._(
-      {Key? key,
-      required this.axis,
-      required this.children,
-      required this.controller,
-      required this.dividerBuilder,
-      required this.onDividerDragUpdate,
-      required this.onDividerTap,
-      required this.onDividerDoubleTap,
-      required this.resizable,
-      required this.antiAliasingWorkaround,
-      required this.count,
-      required this.widgetBuilder,
-      required this.initialAreas,
-      required this.pushDividers,
-      required this.sizeOverflowPolicy,
-      required this.sizeUnderflowPolicy,
-      required this.fallbackWidth,
-      required this.fallbackHeight})
+      this.axis = MultiSplitView.defaultAxis,
+      this.controller,
+      this.dividerBuilder,
+      this.onDividerDragUpdate,
+      this.onDividerTap,
+      this.onDividerDoubleTap,
+      this.resizable = true,
+      this.antiAliasingWorkaround = false,
+      this.pushDividers = false,
+      this.initialAreas,
+      this.sizeOverflowPolicy = SizeOverflowPolicy.shrinkLast,
+      this.sizeUnderflowPolicy = SizeUnderflowPolicy.stretchLast,
+      this.fallbackWidth = 500,
+      this.fallbackHeight = 500,
+      this.builder})
       : super(key: key);
 
   final Axis axis;
-  final List<Widget>? children;
   final MultiSplitViewController? controller;
   final List<Area>? initialAreas;
+
+  final AreaWidgetBuilder? builder;
 
   /// Indicates whether a divider can push others.
   final bool pushDividers;
@@ -142,10 +62,6 @@ class MultiSplitView extends StatefulWidget {
 
   /// Function to listen divider dragging.
   final OnDividerDragUpdate? onDividerDragUpdate;
-
-  final int? count;
-
-  final AreaWidgetBuilder? widgetBuilder;
 
   /// Represents the policy for handling overflow of non-flexible areas within
   /// a container.
@@ -175,13 +91,6 @@ class MultiSplitView extends StatefulWidget {
   ///  * [fallbackWidth], the same but horizontally.
   final double fallbackHeight;
 
-  int get _childrenCount {
-    if (children != null) {
-      return children!.length;
-    }
-    return count ?? 0;
-  }
-
   @override
   State createState() => _MultiSplitViewState();
 }
@@ -197,8 +106,8 @@ class _MultiSplitViewState extends State<MultiSplitView> {
 
   Object? _lastAreasUpdateHash;
 
-  LayoutConstraints _layoutConstraints = LayoutConstraints(
-      childrenCount: 0, containerSize: 0, dividerThickness: 0);
+  LayoutConstraints _layoutConstraints =
+      LayoutConstraints(areasCount: 0, containerSize: 0, dividerThickness: 0);
 
   @override
   void initState() {
@@ -256,132 +165,140 @@ class _MultiSplitViewState extends State<MultiSplitView> {
 
   @override
   Widget build(BuildContext context) {
-    if (widget._childrenCount > 0) {
-      MultiSplitViewThemeData themeData = MultiSplitViewTheme.of(context);
+    MultiSplitViewThemeData themeData = MultiSplitViewTheme.of(context);
 
-      return LayoutBuilder(builder: (context, constraints) {
-        ControllerHelper controllerHelper = ControllerHelper(_controller);
+    return LayoutBuilder(builder: (context, constraints) {
+      ControllerHelper controllerHelper = ControllerHelper(_controller);
 
-        final double containerSize = widget.axis == Axis.horizontal
-            ? constraints.maxWidth
-            : constraints.maxHeight;
+      final double containerSize = widget.axis == Axis.horizontal
+          ? constraints.maxWidth
+          : constraints.maxHeight;
 
-        if (_lastAreasUpdateHash != controllerHelper.areasUpdateHash ||
-            _layoutConstraints.containerSize != containerSize ||
-            _layoutConstraints.childrenCount != widget._childrenCount ||
-            _layoutConstraints.dividerThickness != themeData.dividerThickness) {
-          _draggingDividerIndex = null;
-          _lastAreasUpdateHash = controllerHelper.areasUpdateHash;
+      if (_lastAreasUpdateHash != controllerHelper.areasUpdateHash ||
+          _layoutConstraints.containerSize != containerSize ||
+          _layoutConstraints.areasCount != _controller.areasCount ||
+          _layoutConstraints.dividerThickness != themeData.dividerThickness) {
+        _draggingDividerIndex = null;
+        _lastAreasUpdateHash = controllerHelper.areasUpdateHash;
 
-          _layoutConstraints = LayoutConstraints(
-              childrenCount: widget._childrenCount,
-              containerSize: containerSize,
-              dividerThickness: themeData.dividerThickness);
-          _layoutConstraints.adjustAreas(
-              controllerHelper: controllerHelper,
-              sizeOverflowPolicy: widget.sizeOverflowPolicy,
-              sizeUnderflowPolicy: widget.sizeUnderflowPolicy);
+        _layoutConstraints = LayoutConstraints(
+            areasCount: _controller.areasCount,
+            containerSize: containerSize,
+            dividerThickness: themeData.dividerThickness);
+        _layoutConstraints.adjustAreas(
+            controllerHelper: controllerHelper,
+            sizeOverflowPolicy: widget.sizeOverflowPolicy,
+            sizeUnderflowPolicy: widget.sizeUnderflowPolicy);
+      }
+
+      List<Widget> children = [];
+
+      for (int index = 0; index < _controller.areasCount; index++) {
+        Area area = _controller.getArea(index);
+
+        // area widget
+        Widget child;
+        if (area.widget != null) {
+          child = area.widget!;
+        } else if (area.builder != null) {
+          child = area.builder!(context);
+        } else if (widget.builder != null) {
+          child = widget.builder!(context, index, area);
+        } else {
+          child = Container();
         }
 
-        List<Widget> children = [];
+        child = IgnorePointer(
+            child: child, ignoring: _draggingDividerIndex != null);
 
-        for (int index = 0; index < _controller.areasCount; index++) {
-          // area widget
-          Widget child = widget.children != null
-              ? widget.children![index]
-              : widget.widgetBuilder!(context, index, _controller.areas[index]);
+        MouseCursor cursor = MouseCursor.defer;
+        if (_draggingDividerIndex != null) {
+          cursor = widget.axis == Axis.horizontal
+              ? SystemMouseCursors.resizeColumn
+              : SystemMouseCursors.resizeRow;
+        }
+        child = MouseRegion(cursor: cursor, child: child);
+        children.add(LayoutId(
+            key: AreaHelper.keyFrom(area: area),
+            id: index,
+            child: ClipRect(child: child)));
 
-          child = IgnorePointer(
-              child: child, ignoring: _draggingDividerIndex != null);
-
-          MouseCursor cursor = MouseCursor.defer;
-          if (_draggingDividerIndex != null) {
-            cursor = widget.axis == Axis.horizontal
-                ? SystemMouseCursors.resizeColumn
-                : SystemMouseCursors.resizeRow;
-          }
-          child = MouseRegion(cursor: cursor, child: child);
-          children.add(LayoutId(id: index, child: ClipRect(child: child)));
-
-          //divisor widget
-          if (index < _controller.areasCount - 1) {
-            bool highlighted = (_draggingDividerIndex == index ||
-                (_draggingDividerIndex == null && _hoverDividerIndex == index));
-            Widget dividerWidget = widget.dividerBuilder != null
-                ? widget.dividerBuilder!(
-                    widget.axis == Axis.horizontal
-                        ? Axis.vertical
-                        : Axis.horizontal,
-                    index,
-                    widget.resizable,
-                    _draggingDividerIndex == index,
-                    highlighted,
-                    themeData)
-                : DividerWidget(
-                    axis: widget.axis == Axis.horizontal
-                        ? Axis.vertical
-                        : Axis.horizontal,
-                    index: index,
-                    themeData: themeData,
-                    highlighted: highlighted,
-                    resizable: widget.resizable,
-                    dragging: _draggingDividerIndex == index);
-            if (widget.resizable) {
-              dividerWidget = GestureDetector(
-                  behavior: HitTestBehavior.translucent,
-                  onTap: () => _onDividerTap(index),
-                  onDoubleTap: () => _onDividerDoubleTap(index),
-                  onHorizontalDragDown: widget.axis == Axis.vertical
-                      ? null
-                      : (detail) => _onDragDown(detail, index),
-                  onHorizontalDragCancel: widget.axis == Axis.vertical
-                      ? null
-                      : () => _onDragCancel(),
-                  onHorizontalDragEnd: widget.axis == Axis.vertical
-                      ? null
-                      : (detail) => _onDragEnd(),
-                  onHorizontalDragUpdate: widget.axis == Axis.vertical
-                      ? null
-                      : (detail) =>
-                          _onDragUpdate(detail, index, controllerHelper),
-                  onVerticalDragDown: widget.axis == Axis.horizontal
-                      ? null
-                      : (detail) => _onDragDown(detail, index),
-                  onVerticalDragCancel: widget.axis == Axis.horizontal
-                      ? null
-                      : () => _onDragCancel(),
-                  onVerticalDragEnd: widget.axis == Axis.horizontal
-                      ? null
-                      : (detail) => _onDragEnd(),
-                  onVerticalDragUpdate: widget.axis == Axis.horizontal
-                      ? null
-                      : (detail) =>
-                          _onDragUpdate(detail, index, controllerHelper),
-                  child: dividerWidget);
-              dividerWidget = _mouseRegion(
-                  index: index,
+        //divisor widget
+        if (index < _controller.areasCount - 1) {
+          bool highlighted = (_draggingDividerIndex == index ||
+              (_draggingDividerIndex == null && _hoverDividerIndex == index));
+          Widget dividerWidget = widget.dividerBuilder != null
+              ? widget.dividerBuilder!(
+                  widget.axis == Axis.horizontal
+                      ? Axis.vertical
+                      : Axis.horizontal,
+                  index,
+                  widget.resizable,
+                  _draggingDividerIndex == index,
+                  highlighted,
+                  themeData)
+              : DividerWidget(
                   axis: widget.axis == Axis.horizontal
                       ? Axis.vertical
                       : Axis.horizontal,
-                  dividerWidget: dividerWidget,
-                  themeData: themeData);
-            }
-            children.add(LayoutId(id: 'd$index', child: dividerWidget));
+                  index: index,
+                  themeData: themeData,
+                  highlighted: highlighted,
+                  resizable: widget.resizable,
+                  dragging: _draggingDividerIndex == index);
+          if (widget.resizable) {
+            dividerWidget = GestureDetector(
+                behavior: HitTestBehavior.translucent,
+                onTap: () => _onDividerTap(index),
+                onDoubleTap: () => _onDividerDoubleTap(index),
+                onHorizontalDragDown: widget.axis == Axis.vertical
+                    ? null
+                    : (detail) => _onDragDown(detail, index),
+                onHorizontalDragCancel:
+                    widget.axis == Axis.vertical ? null : () => _onDragCancel(),
+                onHorizontalDragEnd: widget.axis == Axis.vertical
+                    ? null
+                    : (detail) => _onDragEnd(),
+                onHorizontalDragUpdate: widget.axis == Axis.vertical
+                    ? null
+                    : (detail) =>
+                        _onDragUpdate(detail, index, controllerHelper),
+                onVerticalDragDown: widget.axis == Axis.horizontal
+                    ? null
+                    : (detail) => _onDragDown(detail, index),
+                onVerticalDragCancel: widget.axis == Axis.horizontal
+                    ? null
+                    : () => _onDragCancel(),
+                onVerticalDragEnd: widget.axis == Axis.horizontal
+                    ? null
+                    : (detail) => _onDragEnd(),
+                onVerticalDragUpdate: widget.axis == Axis.horizontal
+                    ? null
+                    : (detail) =>
+                        _onDragUpdate(detail, index, controllerHelper),
+                child: dividerWidget);
+            dividerWidget = _mouseRegion(
+                index: index,
+                axis: widget.axis == Axis.horizontal
+                    ? Axis.vertical
+                    : Axis.horizontal,
+                dividerWidget: dividerWidget,
+                themeData: themeData);
           }
+          children.add(LayoutId(id: 'd$index', child: dividerWidget));
         }
-        return LimitedBox(
-            maxWidth: widget.fallbackWidth,
-            maxHeight: widget.fallbackHeight,
-            child: CustomMultiChildLayout(
-                children: children,
-                delegate: LayoutDelegate(
-                    controller: _controller,
-                    axis: widget.axis,
-                    layoutConstraints: _layoutConstraints,
-                    antiAliasingWorkaround: widget.antiAliasingWorkaround)));
-      });
-    }
-    return Container();
+      }
+      return LimitedBox(
+          maxWidth: widget.fallbackWidth,
+          maxHeight: widget.fallbackHeight,
+          child: CustomMultiChildLayout(
+              children: children,
+              delegate: LayoutDelegate(
+                  controller: _controller,
+                  axis: widget.axis,
+                  layoutConstraints: _layoutConstraints,
+                  antiAliasingWorkaround: widget.antiAliasingWorkaround)));
+    });
   }
 
   /// Updates the hover divider index.
@@ -502,3 +419,11 @@ class _MultiSplitViewState extends State<MultiSplitView> {
     return container.globalToLocal(globalPosition);
   }
 }
+
+typedef AreaWidgetBuilder = Widget Function(
+    BuildContext context, int index, Area area);
+
+typedef DividerBuilder = Widget Function(Axis axis, int index, bool resizable,
+    bool dragging, bool highlighted, MultiSplitViewThemeData themeData);
+
+typedef OnDividerDragUpdate = void Function(int index);